+++ src/main/java/handler/RequestHandler.java
@@ -4,7 +4,7 @@ import http.HttpRequest;
 import http.HttpResponse;
 import http.status.HttpStatusCode;
 import util.FileUtils;
-import util.HttpResponseUtils;
+import http.HttpResponseFactory;
 import util.UriParser;
 
 import java.util.Optional;
@@ -16,12 +16,11 @@ public interface RequestHandler {
     default HttpResponse getHttpResponse(String uri){
         Optional<byte[]> body = FileUtils.readFile(uri);
         if (body.isPresent()) {
-			return HttpResponseUtils.getHttpResponse(HttpStatusCode.OK, UriParser.getFileType(uri), body.get());
+			return HttpResponseFactory.getHttpResponse(HttpStatusCode.OK, UriParser.getFileType(uri), body.get());
         }
 		//body가 empty라면 404 페이지 전송
 		body = FileUtils.readFile("/error/404.html");
 		//404 페이지를 찾을 수 없으면 "404 NOT FOUND" 문자열 전송
-		return HttpResponseUtils.getHttpResponse(HttpStatusCode.NOT_FOUND,
-			UriParser.getFileType(uri), body.orElse("404 NOT FOUND".getBytes()));
+		return HttpResponseFactory.getHttpResponse(HttpStatusCode.NOT_FOUND, UriParser.getFileType(uri), body.orElse("404 NOT FOUND".getBytes()));
     }
 }
